Attachment with abstract class Attachment and Array in Post.
